Issue #1058: We must use the 'target' attributes to variables targetted by 'c_loc' in 'gridStateVectorFileIO_mod'.
src/modules/gridStateVectorFileIO_mod.f90(1888): error #9022: The argument to C_LOC must be a variable with the POINTER or TARGET attribute.   [WORK2D_R4]
          fstRecords(0)%data = c_loc(work2d_r4(:,:,0))
-------------------------------------^
src/modules/gridStateVectorFileIO_mod.f90(2145): error #9022: The argument to C_LOC must be a variable with the POINTER or TARGET attribute.   [WORK2DFILE_R4]
      fstRecordTmp%data = c_loc(work2dFile_r4)
--------------------------------^
src/modules/gridStateVectorFileIO_mod.f90(2151): error #9022: The argument to C_LOC must be a variable with the POINTER or TARGET attribute.   [DATA]
      fstRecordTmp%data = c_loc(data)
--------------------------------^
src/modules/gridStateVectorFileIO_mod.f90(2170): error #9022: The argument to C_LOC must be a variable with the POINTER or TARGET attribute.   [MASK]
      fstRecordTmp%data = c_loc(mask)
--------------------------------^