Issue #1033: Bugfix in bMatrixEnsemble_mod and in the units of B column diagnotics
A bug was corrected in bMatrixEnsemble_mod when using GEM-H (this is the same bug that was fixed in v4.1 in #1060 and !965).

It was also found that the units of the results of B columns computed by diagBmatrix were incorrect for many variables. The proper corrections were made and the assignement of the appropriate multiplicative factor for unit conversion was simplified and centralized in two new routines in utilities_mod.