Issue #1065: Adding 'contiguous' keyword for all 'mmpi_*' array arguments
The routines in RPN_COMM are assuming contiguous arrays.  Since arrays
can be sliced, the arrays given as arguments to routines may not be
contiguous.  To guarantee contiguous arrays, we add the keyword
'contiguous' to all routines and the compiler will automatically
create temporary contiguous arrays.

You can see the discussion at
https://gitlab.science.gc.ca/atmospheric-data-assimilation/midas/-/commit/ecec329b45ba05e41244a46c7eb6998fe5a12693#note_1152017

At this point, all the system tests pass!