Issue #1065: Increase just a little the resources of '/Tests/diagBmatrix/Bcolumns-vertSDL'
That test was very very tight on memory.  We asked for 9GB for each
MPI tile.  But the first rank was using 9.567 GB.  By chance, it was
passing.

But, with recent changes is seems that the memory usage has slightly
increased to 9.578 GB and the task aborted.

By increased the memory request to 11 GB for each MPI tiles, the tests
should pass easily each time.  This change does not affect the result
of that test.