Issue #1079: Add the namelist variable 'allReduceOrder' to 'NAMCFN'
This namelist variable can be 'forward' or 'backward' and will control
the way the sum is done in the routine 'mmpi_allreduce_sumreal8scalar'.

A validation is performed to make sure the variable is one of the
expected values.

The namelist '/Tests/var/Var3D/splitobs_ppp' includes that variable in
its namelist with the value 'backward'.  This is not affecting the
results yet because the feature has been been implemented yet.