Issue #1071: Update results for '/Tests/letkf/glb_10km' and '/Tests/letkf/glb_15km'
There was a severe bug in RMNLib when using the 'classic' fst98
interface with RSF files where the validity was ignored when reading a
field.  It was giving the first record matching without checking for
the date.

So the ensemble inputs to the LETKF was all wrong.  This is fixed now.