Issue #1096 Fixing the bug creating abort due to floatting-point precission error
For the certain values of latitudes and longitudes , the computed value of a in
the Haversine formula becomes slightly greater than 1 due to floating-point
precision errors. This causes 1 - a < 0, leading to a sqrt of a negative number,
and results in a NaN. A simple fix is introduced to prevent this issue.