Issue #695: Avoid running 'make install' from 'src/midas_build'
Since 'make install' compiles every program before copying them into
the install directory, there is no point to call 'make install' if we
only compile a subset of the targets.

In that case, we just copies from the build directory to the installed
directory the programs compiled to avoid calling 'make install' which
would triggers a compilation of all the programs.