Issue #1097: Modifications to pressure and altitude computations with GEM-H
In the first implementation, pressure and altitude were computed this way:

Z: vgd_levels -> gZ -> Z
P1: Z -> hyposmetric equation with altitude and latitude-dependent g  -> P

This was problematic since 1) GEM-H computes P directly from GZ obtained from vgd_levels and 2) the formulation to compute Z from GZ was incorrect.

The GZ to Z formulation was modified and P is now computed as in GEM, i.e.:

P2: vgd_levels -> gZ  -> hyposmetric equation constant g -> P

With the new approach, the fit to radiance and GPS-RO observations is very close to the one obtained with GEM-P and the previously observed biases in the temperature increments over the Tropics, peaking at 150 and 50 hPa, are removed.

However, a bug still remains when using a GEM-H analysis grid as the code aborts in RTTOV modules.