Issue #1097: Bugfix in int_vInterp_col
Given that geopotential heights are not converted into altitudes at the diagnostic levels, for some grid point in the mountains, the altitude at NLEV-1 can become smaller than the altitude at NLEV level, leading to non-monotonic profiles.

This problem, in combination with the fact that int_vInterp_col applies the vertical interpolation procedure even when no interpolation is needed led non-realistic profiles in some columns.

The quickfix chosen is simply to use geopotential height to interpolate from GEM-H to GEM-H in int_vInterp_col.

The fundamental issues will be adressed in two other separated issues: 1) to prevent int_vInterp_col from applying the interpolation procedure when it is not needed and 2) to correctly compute altitude at all vertical levels or to adopt geopotential height in stateVectors and columns.