Issue #1097: Apply gz2alt conversion at diagnostic levels with GEM-H
It was discovered that not converting geopotential heights to altitudes was creating problem in the satellite bias correction in the context of GEM-H. This is due to the fact that below the ground interpolations are needed to estimate the height of 1000-hPa level and not converting the geopotential heights at the diagnostic levels was impacting the altitude gradient between nLev-1 and nLev; thus negatively impacting the estimated height of 1000-hPa level.

Note that the code is still potentially problematic since we are using altitude and not geopotential height to estimate the height of pressure levels for the bias correction. However, MIDAS with GEM-P follows the same approach.