Issue #1097: Added a patch in biasCorrectionSat_mod when addHeightSfcOffset is true
When the diagnotic levels are assigned their right heights with addHeightSfcOffset = .true., this changes the resulting heights of the 1000-hPa level computed in biasCorrectionSat_mod and introduce non-neglegibel differences between GEM-P and GEM-H since the GEM-H version works only with addHeightSfcOffset = .true. whereas all the existing applications with GEM-P use addHeightSfcOffset = .false.

Thus, to enforce consistency between GEM-P and GEM-H as much as possible and based on previous results showing that GEM-P with addHeightSfcOffset = .true. led to degradations in the upper levels, it was decided to add a pacth in biasCorrectionSat_mod to use the altitude at the surface for heigth at nLev_T in the computation of the height of pressure levels located below the surface.