Issue #1099: second-round initial implemetation
The previous implementation (commit ids 2c7ea01c 801a4d80 b39e0ba6) was based on a bad design of computing both FSOI and FSR in a single execution.
In this implementation, I changed the design so that a single execution only computes either FSOI or FSR.
Which one to compute is controlled by the namelist variable fsoMode.