gfortran: Protect the code for the non-thread-save OpenBLAS library
In the routine 'enkf_algorithmCVLETKF', a call to a BLAS routine for
eigenvalue decomposition is done in an OpenMP region.

When compiling with 'gfortran', we use the system installation of
'openblas'.  As documented in
http://www.openmathlib.org/OpenBLAS/docs/faq/#using-openblas-with-openmp,
OpenBLAS must be compiled with a special option to allow its use in an
OpenMP context.

It appears that the system installation for
'ORDENV_PLAT=rhel-8-icelake-64' is not correct for OpenMP context (see
the patch at the end which shows how to determine this).  So, we avoid
calling concurrently the BLAS routine 'dsyev' by imposing that only
one thread executes at the same time ('!$omp critical').  With this
change, the test '/Tests/letkf/glbGridSpaceLoc' is working.  I also
tested using 'export OPENBLAS_NUM_THREADS=1' but it didn't work.

To find which options have been used to compile the system
installation of OpenBLAS, we use those changes to call
'openblas_get_parallel' (coming from
http://www.openmathlib.org/OpenBLAS/docs/faq/#how-can-i-find-out-at-runtime-what-options-the-library-was-built-with)

diff --git a/src/modules/enkf_mod.f90 b/src/modules/enkf_mod.f90
index 37a325ad6..fbc09f54b 100644
--- a/src/modules/enkf_mod.f90
+++ b/src/modules/enkf_mod.f90
@@ -99,6 +99,11 @@ module enkf_mod
     integer  :: numFullEns             = 0
   end type struct_enkfNML

+  INTERFACE
+    INTEGER FUNCTION openblas_get_parallel() BIND(C, NAME='openblas_get_parallel')
+    END FUNCTION openblas_get_parallel
+  END INTERFACE
+
 contains

   !----------------------------------------------------------------------
@@ -1498,6 +1503,8 @@ contains
     call utl_eigenDecomp(YbTinvRYb_mean, eigenValues_mean, eigenVectors_mean, tolerance, matrixRank)
     call utl_tmg_stop(135)

+    write(*,*) 'enkf_algorithmCVLETKFME: openblas_get_parallel = ', openblas_get_parallel()
+
     ! Compute ensemble mean local weights as E * (Lambda + (Nens-1)*I)^-1 * E^T * YbTinvR * (obs - meanYb)
     weightsTemp(:) = 0.0d0
     do localObsIndex = 1, numLocalObs

Or use this C program:
        #include <stdio.h>
        #include <cblas.h> // Or <cblas.h> depending on your OpenBLAS setup

        int main() {
            int parallel_type = openblas_get_parallel();

            printf("OpenBLAS config: %s\n", openblas_get_config());

            printf("OpenBLAS parallelization type: %d\n", parallel_type);

            if (parallel_type == 0) {
                printf("OpenBLAS is running in single-threaded mode.\n");
            } else if (parallel_type == 1) {
                printf("OpenBLAS is using a non-OpenMP parallelization (e.g., pthreads).\n");
            } else if (parallel_type == 2) {
                printf("OpenBLAS is using OpenMP for parallelization.\n");
            } else {
                printf("Unknown OpenBLAS parallelization type.\n");
            }

            return 0;
        }
which gives this result:
        OpenBLAS config: OpenBLAS 0.3.12 DYNAMIC_ARCH NO_AFFINITY Prescott SINGLE_THREADED
        OpenBLAS parallelization type: 0
        OpenBLAS is running in single-threaded mode.

This program was compiled with
        gcc check_openblas_config.c -o check_openblas_config -lopenblas -I/usr/include/openblas