Fix a compiler warning about output format in routine 'calcVertModesSpec' of module 'calcStatsGlb_mod'
The results of the test '/Tests/calcStats/vertModes_glb' myst be updated
because this changed the output of some text files but not the content
itself.

src/modules/calcStatsGlb_mod.f90(3458): remark #8291: Recommended relationship between field width 'W' and the number of fractional digits 'D' in this edit descriptor is 'W>=D+7'.
          write(99,'(I4,2X,E10.4)') modeIndex, powerSpec(varIndex,modeIndex)
----------------------------^