Issue #1100: Fix a bug in 'maestro/suites/midas_system_tests/modules/UnitTest/run.tsk'
The regex in 'run.tsk' to identify the correct timing statistics was
not distinguishing test starting with the same name like
'/Tests/oMinusF/o-p' and '/Tests/oMinusF/o-p_chem'.

Because of that bug, the statistics for the test '/Tests/oMinusF/o-p'
and '/Tests/obsSelection/conv' were empty.  So they have been updated
with the statistics on 'ppp8' and 'conrad' that ran in parallel.