Issue #1100: Fix compilation errors from Intel 2025 in 'src/modules/gridStateVectorFileIO_mod.f90'
[ 33%] Building Fortran object src/modules/CMakeFiles/midas.dir/gridStateVectorFileIO_mod.f90.o
src/modules/gridStateVectorFileIO_mod.f90(1891): error #9022: The argument to C_LOC must be a variable with the POINTER or TARGET attribute.   [WORK2D_R4]
          fstRecords(0)%data = c_loc(work2d_r4(:,:,0))
-------------------------------------^
src/modules/gridStateVectorFileIO_mod.f90(1945): error #9022: The argument to C_LOC must be a variable with the POINTER or TARGET attribute.   [WORK2D_R4]
            fstRecords(0)%data = c_loc(work2d_r4(:,:,0))
---------------------------------------^
src/modules/gridStateVectorFileIO_mod.f90(2195): error #9022: The argument to C_LOC must be a variable with the POINTER or TARGET attribute.   [WORK2DFILE_R4]
      fstRecordTmp%data = c_loc(work2dFile_r4)
--------------------------------^
src/modules/gridStateVectorFileIO_mod.f90(2201): error #9022: The argument to C_LOC must be a variable with the POINTER or TARGET attribute.   [DATA]
      fstRecordTmp%data = c_loc(data)
--------------------------------^
src/modules/gridStateVectorFileIO_mod.f90(2220): error #9022: The argument to C_LOC must be a variable with the POINTER or TARGET attribute.   [MASK]
      fstRecordTmp%data = c_loc(mask)
--------------------------------^