Issue #1100: Adjust the reference results for some tests on HPCR-U3
The results of the following tests
   * /Tests/oMinusF/o-p
   * /Tests/oMinusF/o-a
   * /Tests/obsSelection/BgckCris
   * /Tests/obsSelection/BgckCsr
   * /Tests/obsSelection/bgckIasiwBcor

are not affected by code changes between the branches 'v_3.9',
'v_3.10' and 'v_4.1' on HPCR-U2.  But they are affected by the same
code changes when run on HPCR-U3.  Some logic was then added in the
configuration files for those tests to select the correct version.